Move callbacks interface to a separate file
